Refactor download functionality into NativeDownloadChannel for improved structure and maintainability
